// frontend/src/app/core/api/model/enrollment-search-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EnrollmentSearchRequest { 
    keyword?: string;
    levelId?: number;
    specialtyId?: number;
    promotionId?: number;
    status?: EnrollmentSearchRequest.StatusEnum;
    page?: number;
    size?: number;
}
export namespace EnrollmentSearchRequest {
    export const StatusEnum = {
        PreEnrolled: 'PRE_ENROLLED',
        Incomplete: 'INCOMPLETE',
        Waitlisted: 'WAITLISTED',
        ConditionallyValidated: 'CONDITIONALLY_VALIDATED',
        Validated: 'VALIDATED',
        Suspended: 'SUSPENDED',
        DroppedOut: 'DROPPED_OUT',
        Rejected: 'REJECTED',
        Cancelled: 'CANCELLED',
        Completed: 'COMPLETED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}